Rename log.py to logger.py to prevent name clashes
